export: Move source files to internal directory
