+++ src/export/export.pro
@@ -8,12 +8,12 @@ CONFIG += staticlib
 
 SOURCES += \
     ExportModule.cpp \
-    ExportProvider.cpp \
-    ExportTable.cpp \
-    ExportTableModel.cpp
+    internal/ExportProvider.cpp \
+    internal/ExportTable.cpp \
+    internal/ExportTableModel.cpp
 
 HEADERS += \
     ExportModule.h \
-    ExportProvider.h \
-    ExportTable.h \
-    ExportTableModel.h
+    internal/ExportProvider.h \
+    internal/ExportTable.h \
+    internal/ExportTableModel.h
